BG DataTable 지역테이블에 날씨정보 추가

+++ ReBloom/Assets/Scripts/Stage/StageDB.cs
@@ -30,15 +30,15 @@ public class StageDB
 
         data.id = entity.Get<int>("Stage_ID");
         data.name = entity.Get<string>("Stage_Name");
-        data.stageVariation = entity.Get<int>("Stage_Variation");
+        data.stageVariation = entity.Get<float>("Stage_Variation");
         data.stagePollution = entity.Get<float>("Stage_Pollution");
         data.stageTemp = entity.Get<float>("Stage_Temp");
         data.sunnyRate = entity.Get<float>("Sunny_Rate");
-        data.sunny_d = entity.Get<int>("Sunny_D");
-        data.sunny_vari = entity.Get<int>("Sunny_Vari");
+        data.sunny_d = entity.Get<float>("Sunny_D");
+        data.sunny_vari = entity.Get<float>("Sunny_Vari");
         data.sunnyPollution = entity.Get<float>("Sunny_Pollution");
         data.sunnyThirst = entity.Get<float>("Sunny_Thirst");
-        data.sunnyTemp = entity.Get<int>("Sunny_Temp");
+        data.sunnyTemp = entity.Get<float>("Sunny_Temp");
 
         return data;
     }
